Add chart time controls Add RWZI selector Fix obile legend render
