+++ src/pages/gemeente/[code]/rioolwater.tsx
@@ -6,7 +6,7 @@ import { SewerWaterBarScale } from '~/components/gemeente/sewer-water-barscale';
 
 import RioolwaterMonitoring from '~/assets/rioolwater-monitoring.svg';
 
-import { RegionalSewerWaterLineChart } from '~/components/lineChart/regionalSewerWaterLineChart';
+import { MunicipalSewerWaterLineChart } from '~/components/lineChart/municipalSewerWaterLineChart';
 import { useMemo } from 'react';
 import { BarChart } from '~/components/charts';
 import {
@@ -69,7 +69,7 @@ const SewerWater: FCWithLayout<IMunicipalityData> = (props) => {
         <h3>{text.linechart_titel}</h3>
 
         {lineChartData && (
-          <RegionalSewerWaterLineChart
+          <MunicipalSewerWaterLineChart
             averageValues={lineChartData.averageValues}
             text={{
               average_label_text: lineChartData.averageLabelText,
