+++ src/scss/layout.scss
@@ -74,6 +74,23 @@
   }
 }
 
+.metric-article-header {
+  display: flex;
+  flex-direction: row;
+  align-items: flex-start;
+
+  div {
+    margin-left: auto;
+  }
+
+  @media (max-width: 1000px) {
+    flex-direction: column;
+    div {
+      margin-left: 0;
+    }
+  }
+}
+
 .index-article,
 .metric-article {
   background: white;
